debugging necessary: day 3, part 2
There seems to be an off-by-one error somewhere, but I cannot figure it
out yet. It is even possible that this error is already hidden in day 3,
part 1, but does not affect the result because the missing visited house
has already been visited earlier (and now, in part 2, this is no longer
the case).
This will be a hell of a debug process...