solve: 2016, day 6, part 1 & 2
Did some HashMap stuff on CodingBat (this would have simplified day 5 as
well...), and I'm especially proud that I needed only SIX MINUTES to get
my solution for part 1 ‘refactored’ to also produce the correct result
for part 2 (including a test run on the given test data).

Maybe I'm not as awful a programmer as I sometimes think I am. It's
mainly a lack of practice. Years of practice, however.